convert finetuning dataset
