+++ llmfoundry/data_prep/__init__.py
@@ -1,0 +1,11 @@
+from llmfoundry.data_prep.convert_finetuning_dataset import (
+    convert_finetuning_dataset,
+    convert_finetuning_dataset_from_args,
+    get_columns_and_format,
+)
+
+__all__ = [
+    'convert_finetuning_dataset',
+    'convert_finetuning_dataset_from_args',
+    'get_columns_and_format',
+]
